+++ nengo_dl/benchmarks.py
@@ -156,8 +156,7 @@ def compare_backends(raw=False):
                                 continue
 
                             if backend == nengo_dl:
-                                kwargs = {"step_blocks": 50,
-                                          "unroll_simulation": True,
+                                kwargs = {"unroll_simulation": 50,
                                           "minibatch_size": 1,
                                           "device": "/gpu:0"
                                           }
@@ -218,9 +217,9 @@ def profiling():
     # note: in order for GPU profiling to work, you have to manually add
     # ...\CUDA\v8.0\extras\CUPTI\libx64 to your path
     net = pes(128, 32, nengo.RectifiedLinear())
-    with nengo_dl.Simulator(net, tensorboard=False, step_blocks=50,
-                            device="/gpu:0", unroll_simulation=True) as sim:
-        sim.run_steps(50, profile=True)
+    with nengo_dl.Simulator(net, tensorboard=False, unroll_simulation=50,
+                            device="/gpu:0") as sim:
+        sim.run_steps(150, profile=True)
 
 
 if __name__ == "__main__":