+++ nengo_dl/simulator.py
@@ -1,4 +1,4 @@
-from __future__ import print_function
+from __future__ import print_function, division
 
 from collections import Mapping
 import datetime
@@ -41,14 +41,10 @@ class Simulator(object):
     device : None or ``"/cpu:0"`` or ``"/gpu:[0-n]"``, optional
         device on which to execute computations (if None then uses the
         default device as determined by Tensorflow)
-    step_blocks : int, optional
-        controls how many simulation steps run each time the graph is
-        executed (affects memory usage and graph construction time)
-    unroll_simulation : bool, optional
-        if True, unroll simulation loop by explicitly building each iteration
-        (up to ``step_blocks``) into the computation graph. if False, use a
-        symbolic loop, which is more general and produces a simpler graph, but
-        is likely to be slower to simulate
+    unroll_simulation : int, optional
+        unroll simulation loop by explicitly building the given number of
+        iterations into the computation graph (improves simulation speed
+        but increases build time)
     minibatch_size : int, optional
         the number of simultaneous inputs that will be passed through the
         network
@@ -76,23 +72,28 @@ class Simulator(object):
          "tests/test_nengo_tests.py:test_args"),
 
         ("nengo/tests/test_node.py:test_unconnected_node",
-         "need to set `step_blocks` to ensure node runs the correct number "
-         "of times (see tests/test_nengo_tests.py:test_unconnected_node"),
+         "need to set `unroll_simulation` to ensure node runs the correct "
+         "number of times (see "
+         "tests/test_nengo_tests.py:test_unconnected_node"),
     ]
 
     def __init__(self, network, dt=0.001, seed=None, model=None,
-                 dtype=tf.float32, device=None, step_blocks=None,
-                 unroll_simulation=False, minibatch_size=None,
-                 tensorboard=False):
+                 dtype=tf.float32, device=None, unroll_simulation=1,
+                 minibatch_size=None, tensorboard=False,
+                 step_blocks="deprecated"):
         self.closed = None
         self.sess = None
         self.tensorboard = tensorboard
-        self.step_blocks = step_blocks
+        self.unroll = unroll_simulation
         self.minibatch_size = 1 if minibatch_size is None else minibatch_size
 
-        if unroll_simulation and step_blocks is None:
-            raise SimulationError("`step_blocks` must be specified when the "
-                                  "simulation is being unrolled")
+        if step_blocks != "deprecated" or isinstance(unroll_simulation, bool):
+            # TODO: remove this in 0.5
+            warnings.warn(
+                "`step_blocks` has been deprecated and will be ignored; "
+                "`Simulator(..., unroll_simulation=n)` is now equivalent to "
+                "`Simulator(..., unroll_simulation=True, step_blocks=n).",
+                DeprecationWarning)
 
         # TODO: allow the simulator to be called flexibly with/without
         # minibatching
@@ -116,8 +117,8 @@ class Simulator(object):
 
         # set up tensorflow graph plan
         self.tensor_graph = TensorGraph(
-            self.model, self.dt, step_blocks, unroll_simulation, dtype,
-            self.minibatch_size, device)
+            self.model, self.dt, unroll_simulation, dtype, self.minibatch_size,
+            device)
 
         self.data = ProbeDict(
             self.model.params,
@@ -176,10 +177,16 @@ class Simulator(object):
         # start session
         # note: we need to allow soft placement when using tf.while_loop,
         # because tensorflow pins loop variables to the CPU
+        # TODO: switch allow_soft_placement to False once tensorflow
+        # adds the RefExit GPU kernel
         config = tf.ConfigProto(
-            allow_soft_placement=not self.tensor_graph.unroll_simulation,
+            allow_soft_placement=True,
             log_device_placement=False,
         )
+        # TODO: XLA compiling doesn't seem to provide any benefit at the
+        # moment, revisit later after tensorflow has developed it further
+        # config.graph_options.optimizer_options.global_jit_level = (
+        #     tf.OptimizerOptions.ON_1)
 
         self.sess = tf.Session(graph=self.tensor_graph.graph, config=config)
         self.closed = False
@@ -257,22 +264,22 @@ class Simulator(object):
 
         Notes
         -----
-        If ``step_blocks`` is specified, and ``n_steps > step_blocks``, this
-        will repeatedly execute ``step_blocks`` timesteps until the the number
-        of steps executed is >= ``n_steps``.
+        If ``unroll_simulation=x`` is specified, and ``n_steps > x``, this will
+        repeatedly execute ``x`` timesteps until the the number of steps
+        executed is >= ``n_steps``.
         """
 
         if self.closed:
             raise SimulatorClosed("Simulator cannot run because it is closed.")
 
-        if self.step_blocks is not None and n_steps % self.step_blocks != 0:
+        actual_steps = self.unroll * int(np.ceil(n_steps / self.unroll))
+
+        if actual_steps != n_steps:
             warnings.warn(
-                "Number of steps (%d) is not an even multiple of `step_blocks`"
-                " (%d).  Simulation will run for %d steps, which may have "
-                "unintended side effects." %
-                (n_steps, self.step_blocks,
-                 self.step_blocks * (n_steps // self.step_blocks + 1)),
-                RuntimeWarning)
+                "Number of steps (%d) is not an even multiple of "
+                "`unroll_simulation` (%d).  Simulation will run for %d steps, "
+                "which may have unintended side effects." %
+                (n_steps, self.unroll, actual_steps), RuntimeWarning)
 
         if input_feeds is not None:
             self._check_data(input_feeds, mode="out", check_mini=True,
@@ -280,65 +287,6 @@ class Simulator(object):
 
         print_and_flush("Simulation started", end="")
         start = time.time()
-
-        if self.step_blocks is None:
-            probe_data = self._run_steps(n_steps, input_feeds=input_feeds,
-                                         profile=profile)
-
-            self._update_probe_data(probe_data, self.n_steps - n_steps,
-                                    n_steps)
-        else:
-            # break the run up into `step_blocks` sized chunks
-            count = 0
-            while count < n_steps:
-                sub_feed = (
-                    None if input_feeds is None else
-                    {n: v[:, count:count + self.step_blocks]
-                     for n, v in input_feeds.items()})
-                probe_data = self._run_steps(
-                    self.step_blocks, input_feeds=sub_feed,
-                    profile=profile if count == 0 else False)
-                count += self.step_blocks
-
-                self._update_probe_data(
-                    probe_data, self.n_steps - self.step_blocks,
-                    self.step_blocks + min(n_steps - count, 0))
-
-            # update n_steps/time
-            self.n_steps += n_steps - count
-            self.time = self.n_steps * self.dt
-
-        print("\rSimulation completed in %s" %
-              datetime.timedelta(seconds=int(time.time() - start)))
-
-    def _run_steps(self, n_steps, input_feeds=None, profile=False):
-        """Execute ``step_blocks`` sized segments of the simulation.
-
-        Parameters
-        ----------
-        n_steps : int
-            the number of simulation steps to be executed
-        input_feeds : dict of {:class:`~nengo:nengo.Node`: \
-                               :class:`~numpy:numpy.ndarray`}
-            override the values of input Nodes with the given data.  arrays
-            should have shape ``(sim.minibatch_size, n_steps, node.size_out)``.
-        profile : bool, optional
-            if True, collect TensorFlow profiling information while the
-            simulation is running (this will slow down the simulation)
-
-        Notes
-        -----
-        - This function should not be called directly; run the simulator
-          through :meth:`.Simulator.step`, :meth:`.Simulator.run_steps`, or
-          :meth:`.Simulator.run`.
-
-        - The ``input_feeds`` argument allows the user to pass several
-          simultaneous input sequences through the model.  That is, instead of
-          running the model ``n`` times with 1 input at a time, the model
-          can be run once with ``n`` inputs at a time.  Only the values of
-          input nodes (nodes with no incoming Connections) can be overwritten
-          in this way.
-        """
 
         if profile:
             run_options = tf.RunOptions(trace_level=tf.RunOptions.FULL_TRACE)
@@ -351,20 +299,26 @@ class Simulator(object):
         try:
             steps_run, probe_data = self.sess.run(
                 [self.tensor_graph.steps_run, self.tensor_graph.probe_arrays],
-                feed_dict=self._fill_feed(n_steps, input_feeds,
+                feed_dict=self._fill_feed(actual_steps, input_feeds,
                                           start=self.n_steps),
                 options=run_options, run_metadata=run_metadata)
         except (tf.errors.InternalError, tf.errors.UnknownError) as e:
             if e.op.type == "PyFunc":
                 raise SimulationError(
-                    "Function '%s' caused an error "
-                    "(see error log above)" % e.op.name)
+                    "Function '%s' caused an error (see error log above)" %
+                    e.op.name)
             else:
-                raise e
+                raise e  # pragma: no cover
+
+        # update probe data
+        self._update_probe_data(probe_data, self.n_steps, n_steps)
 
         # update n_steps
-        assert steps_run == n_steps
-        self.n_steps += steps_run
+        # note: we update n_steps according to the number of steps that the
+        # user asked for, not the number of steps that were actually run (
+        # in the case of uneven unroll_simulation)
+        assert steps_run == actual_steps
+        self.n_steps += n_steps
         self.time = self.n_steps * self.dt
 
         if profile:
@@ -372,7 +326,8 @@ class Simulator(object):
             with open("%s/nengo_dl_profile.json" % DATA_DIR, "w") as f:
                 f.write(timeline.generate_chrome_trace_format())
 
-        return probe_data
+        print("\rSimulation completed in %s" %
+              datetime.timedelta(seconds=int(time.time() - start)))
 
     def train(self, inputs, targets, optimizer, n_epochs=1, objective="mse",
               shuffle=True):
@@ -385,12 +340,12 @@ class Simulator(object):
         inputs : dict of {:class:`~nengo:nengo.Node`: \
                           :class:`~numpy:numpy.ndarray`}
             input values for Nodes in the network; arrays should have shape
-            ``(batch_size, sim.step_blocks, node.size_out)``
+            ``(batch_size, n_steps, node.size_out)``
         targets : dict of {:class:`~nengo:nengo.Probe`: \
                            :class:`~numpy:numpy.ndarray`}
             desired output value at Probes, corresponding to each value in
             ``inputs``; arrays should have shape
-            ``(batch_size, sim.step_blocks, probe.size_in)``
+            ``(batch_size, n_steps, probe.size_in)``
         optimizer : ``tf.train.Optimizer``
             Tensorflow optimizer, e.g.
             ``tf.train.GradientDescentOptimizer(learning_rate=0.1)``
@@ -429,16 +384,14 @@ class Simulator(object):
           on Tensorflow GPU support.
         """
 
+        n_steps = next(iter(inputs.values())).shape[1]
+
         # error checking
         if self.closed:
             raise SimulatorClosed("Simulator cannot be trained because it is "
                                   "closed.")
-        if self.step_blocks is None:
-            raise SimulationError(
-                "Simulator `step_blocks` must be set for training "
-                "(`Simulator(..., step_blocks=n)`)")
-        self._check_data(inputs, mode="out")
-        self._check_data(targets, mode="in")
+        self._check_data(inputs, mode="out", n_steps=n_steps)
+        self._check_data(targets, mode="in", n_steps=n_steps)
 
         # check for non-differentiable elements in graph
         # utils.find_non_differentiable(
@@ -463,7 +416,7 @@ class Simulator(object):
                 self.soft_reset()
                 try:
                     self.sess.run([opt_op], feed_dict=self._fill_feed(
-                        self.step_blocks, inp, tar))
+                        n_steps, inp, tar))
                 except tf.errors.InvalidArgumentError:
                     raise SimulationError(
                         "TensorFlow does not yet support this optimizer on "
@@ -480,12 +433,12 @@ class Simulator(object):
         inputs : dict of {:class:`~nengo:nengo.Node`: \
                           :class:`~numpy:numpy.ndarray`}
             input values for Nodes in the network; arrays should have shape
-            ``(batch_size, sim.step_blocks, node.size_out)``
+            ``(batch_size, n_steps, node.size_out)``
         targets : dict of {:class:`~nengo:nengo.Probe`: \
                            :class:`~numpy:numpy.ndarray`}
             desired output value at Probes, corresponding to each value in
             ``inputs``; arrays should have shape
-            ``(batch_size, sim.step_blocks, probe.size_in)``
+            ``(batch_size, n_steps, probe.size_in)``
         objective : ``"mse"`` or callable
             the objective used to compute loss. passing ``"mse"`` will use
             mean squared error. a custom function
@@ -500,16 +453,14 @@ class Simulator(object):
         should not be intermixed with calls to :meth:`.Simulator.run`.
         """
 
+        n_steps = next(iter(inputs.values())).shape[1]
+
         # error checking
         if self.closed:
             raise SimulatorClosed("Loss cannot be computed after simulator is "
                                   "closed.")
-        if self.step_blocks is None:
-            raise SimulationError(
-                "Simulator `step_blocks` must be set to compute loss "
-                "(`Simulator(..., step_blocks=n)`)")
-        self._check_data(inputs, mode="out")
-        self._check_data(targets, mode="in")
+        self._check_data(inputs, mode="out", n_steps=n_steps)
+        self._check_data(targets, mode="in", n_steps=n_steps)
 
         # get loss op
         loss = self.tensor_graph.build_loss(objective, tuple(targets.keys()))
@@ -520,7 +471,7 @@ class Simulator(object):
                 inputs, targets, self.minibatch_size, rng=self.rng)):
             self.soft_reset()
             loss_val += self.sess.run(
-                loss, feed_dict=self._fill_feed(self.step_blocks, inp, tar))
+                loss, feed_dict=self._fill_feed(n_steps, inp, tar))
         self.soft_reset()
         loss_val /= i + 1
 
@@ -653,7 +604,7 @@ class Simulator(object):
             the number of timesteps over which we want to collect data
         """
 
-        # first, remove any extra timesteps (due to `step_blocks` mismatch)
+        # remove any extra timesteps (due to `unroll_simulation` mismatch)
         probe_data = [p[:n_steps] for p in probe_data]
 
         for i, p in enumerate(self.model.probes):
@@ -813,12 +764,12 @@ class Simulator(object):
         """
 
         delta = 1e-3
+        n_steps = self.unroll * 2
 
         feed = self._fill_feed(
-            self.step_blocks, {n: np.zeros((self.minibatch_size,
-                                            self.step_blocks, n.size_out))
-                               for n in self.tensor_graph.invariant_inputs},
-            {p: np.zeros((self.minibatch_size, self.step_blocks, p.size_in))
+            n_steps, {n: np.zeros((self.minibatch_size, n_steps, n.size_out))
+                      for n in self.tensor_graph.invariant_inputs},
+            {p: np.zeros((self.minibatch_size, n_steps, p.size_in))
              for p in self.tensor_graph.target_phs})
 
         if outputs is None:
@@ -831,11 +782,13 @@ class Simulator(object):
         # check gradient wrt inp
         for node, inp in self.tensor_graph.invariant_ph.items():
             inp_shape = inp.get_shape().as_list()
+            inp_shape = [n_steps if x is None else x for x in inp_shape]
             inp_tens = self.tensor_graph.invariant_ph[node]
             feed[inp_tens] = np.ascontiguousarray(feed[inp_tens])
             inp_val = np.ravel(feed[inp_tens])
             for out in outputs:
                 out_shape = out.get_shape().as_list()
+                out_shape = [n_steps if x is None else x for x in out_shape]
 
                 # we need to compute the numeric jacobian manually, to
                 # correctly handle variables (tensorflow doesn't expect
@@ -894,12 +847,12 @@ class Simulator(object):
         check_mini : bool, optional
             if True, also validate minibatch_size
         n_steps : int, optional
-            number of simulation steps (if None, will use sim.step_blocks)
+            number of simulation steps (if None, will use sim.unroll)
         """
 
         for n, x in data.items():
             shape = (self.minibatch_size if check_mini else None,
-                     self.step_blocks if n_steps is None else n_steps,
+                     self.unroll if n_steps is None else n_steps,
                      n.size_out if mode == "out" else n.size_in)
 
             if any(x is not None and x != y for x, y in zip(shape, x.shape)):
@@ -908,7 +861,7 @@ class Simulator(object):
                     "(%s, %s, %s.size_%s) -> %s" %
                     (x.shape,
                      "sim.minibatch_size" if check_mini else "_",
-                     "sim.step_blocks" if n_steps is None else "n_steps",
+                     "sim.unroll" if n_steps is None else "n_steps",
                      n, mode, shape))
 
 
@@ -926,7 +879,7 @@ class ProbeDict(Mapping):
     raw : dict of {:class:`~nengo:nengo.Probe`: \
                    list of :class:`~numpy:numpy.ndarray`}
         the raw probe output from the simulator (a list of arrays containing
-        the output from each ``step_blocks`` execution segment)
+        the output from each ``run_steps`` execution segment)
     minibatches : dict of {:class:`~nengo:nengo.Probe`: int or None}
         the minibatch size for each probe in the dictionary (or -1 if the
         probed signal does not have a minibatch dimension)
@@ -940,28 +893,27 @@ class ProbeDict(Mapping):
     def __init__(self, raw, minibatches):
         self.raw = raw
         self.minibatches = minibatches
-        self._cache = {}
+
+        # TODO: add cache back in?
 
     def __getitem__(self, key):
-        cache_miss = (key not in self._cache or
-                      len(self._cache[key]) != len(self.raw[key]))
-        if cache_miss:
-            rval = self.raw[key]
-            if isinstance(rval, list):
-                # combine data from _run_steps iterations
-                rval = np.concatenate(rval, axis=0)
+        rval = self.raw[key]
 
-                if self.minibatches[key] != -1:
-                    if self.minibatches[key] is None:
-                        # get rid of batch dimension
-                        rval = rval[..., 0]
-                    else:
-                        # move batch dimension to front
-                        rval = np.moveaxis(rval, -1, 0)
+        if isinstance(rval, list):
+            # combine data from run_steps iterations
+            rval = np.concatenate(rval, axis=0)
 
-                rval.setflags(write=False)
-            self._cache[key] = rval
-        return self._cache[key]
+            if self.minibatches[key] != -1:
+                if self.minibatches[key] is None:
+                    # get rid of batch dimension
+                    rval = rval[..., 0]
+                else:
+                    # move batch dimension to front
+                    rval = np.moveaxis(rval, -1, 0)
+
+            rval.setflags(write=False)
+
+        return rval
 
     def __iter__(self):
         return iter(self.raw)