+++ nengo_dl/tests/test_simulator.py
@@ -1,5 +1,4 @@
 from collections import OrderedDict
-import itertools
 import os
 import shutil
 
@@ -22,7 +21,7 @@ def test_persistent_state(Simulator, seed):
         nengo.Connection(inp, ens)
         p = nengo.Probe(ens)
 
-    with Simulator(net, step_blocks=5) as sim:
+    with Simulator(net) as sim:
         sim.run_steps(100)
         data = sim.data[p]
         sim.reset()
@@ -31,8 +30,8 @@ def test_persistent_state(Simulator, seed):
         data2 = sim.data[p]
         sim.reset()
 
-        for _ in range(20):
-            sim.run_steps(5)
+        for _ in range(100 // sim.unroll):
+            sim.run_steps(sim.unroll)
         data3 = sim.data[p]
 
     assert np.allclose(data, data2)
@@ -46,41 +45,16 @@ def test_step_blocks(Simulator, seed):
         nengo.Connection(inp, ens)
         p = nengo.Probe(ens)
 
-    with Simulator(net, step_blocks=25) as sim1:
+    with Simulator(net, unroll_simulation=25) as sim1:
         sim1.run_steps(50)
-    with Simulator(net, step_blocks=10) as sim2:
+    with Simulator(net, unroll_simulation=10) as sim2:
         sim2.run_steps(50)
-    with Simulator(net, unroll_simulation=False, step_blocks=None) as sim3:
-        sim3.run_steps(50)
 
     assert np.allclose(sim1.data[p], sim2.data[p])
-    assert np.allclose(sim2.data[p], sim3.data[p])
 
     with pytest.warns(RuntimeWarning):
-        with Simulator(net, step_blocks=5) as sim:
+        with Simulator(net, unroll_simulation=5) as sim:
             sim.run_steps(2)
-
-
-def test_unroll_simulation(Simulator, seed):
-    # note: we run this multiple times because the effects of unrolling can
-    # be somewhat stochastic depending on the op order
-    for i in range(5):
-        with nengo.Network(seed=seed + i) as net:
-            inp = nengo.Node(np.sin)
-            ens = nengo.Ensemble(10, 1)
-            nengo.Connection(inp, ens)
-            p = nengo.Probe(ens)
-
-        with Simulator(net, step_blocks=None, unroll_simulation=False) as sim1:
-            sim1.run_steps(30)
-
-        with Simulator(net, step_blocks=10, unroll_simulation=True) as sim2:
-            sim2.run_steps(30)
-
-        assert np.allclose(sim1.data[p], sim2.data[p])
-
-    with pytest.raises(SimulationError):
-        Simulator(net, step_blocks=None, unroll_simulation=True)
 
 
 def test_minibatch(Simulator, seed):
@@ -125,7 +99,6 @@ def test_minibatch(Simulator, seed):
 
 def test_input_feeds(Simulator):
     minibatch_size = 10
-    step_blocks = 5
 
     with nengo.Network() as net:
         inp = nengo.Node([0, 0, 0])
@@ -133,30 +106,28 @@ def test_input_feeds(Simulator):
         nengo.Connection(inp, out, synapse=None)
         p = nengo.Probe(out)
 
-    with Simulator(net, minibatch_size=minibatch_size,
-                   step_blocks=step_blocks) as sim:
-        val = np.random.randn(minibatch_size, step_blocks * 4, 3)
-        sim.run_steps(step_blocks * 4, input_feeds={inp: val})
+    with Simulator(net, minibatch_size=minibatch_size) as sim:
+        val = np.random.randn(minibatch_size, 50, 3)
+        sim.run_steps(50, input_feeds={inp: val})
         assert np.allclose(sim.data[p], val)
 
         with pytest.raises(nengo.exceptions.SimulationError):
-            sim.run_steps(step_blocks, input_feeds={
+            sim.run_steps(5, input_feeds={
                 inp: np.random.randn(
-                    minibatch_size + 1, step_blocks, 3)})
+                    minibatch_size + 1, 5, 3)})
 
         with pytest.raises(nengo.exceptions.SimulationError):
-            sim.run_steps(step_blocks, input_feeds={
-                inp: np.random.randn(minibatch_size, step_blocks - 1, 3)})
+            sim.run_steps(5, input_feeds={
+                inp: np.random.randn(minibatch_size, 4, 3)})
 
         with pytest.raises(nengo.exceptions.SimulationError):
-            sim.run_steps(step_blocks, input_feeds={
-                inp: np.random.randn(minibatch_size, step_blocks, 4)})
+            sim.run_steps(5, input_feeds={
+                inp: np.random.randn(minibatch_size, 5, 4)})
 
 
 @pytest.mark.parametrize("neurons", (True, False))
 def test_train_ff(Simulator, neurons, seed):
     minibatch_size = 4
-    step_blocks = 1
     n_hidden = 20
 
     np.random.seed(seed)
@@ -180,7 +151,7 @@ def test_train_ff(Simulator, neurons, seed):
         # TODO: why does training fail if we probe out instead of out.neurons?
         p = nengo.Probe(out.neurons)
 
-    with Simulator(net, minibatch_size=minibatch_size, step_blocks=step_blocks,
+    with Simulator(net, minibatch_size=minibatch_size, unroll_simulation=1,
                    seed=seed) as sim:
         x = np.asarray([[[0, 0]], [[0, 1]], [[1, 0]], [[1, 1]]])
         y = np.asarray([[[0.1]], [[0.9]], [[0.9]], [[0.1]]])
@@ -198,8 +169,8 @@ def test_train_ff(Simulator, neurons, seed):
 def test_train_recurrent(Simulator, seed):
     batch_size = 100
     minibatch_size = 100
-    step_blocks = 10
     n_hidden = 30
+    n_steps = 10
 
     with nengo.Network(seed=seed) as net:
         inp = nengo.Node([0])
@@ -214,28 +185,27 @@ def test_train_recurrent(Simulator, seed):
 
         p = nengo.Probe(out)
 
-    with Simulator(net, minibatch_size=minibatch_size, step_blocks=step_blocks,
-                   seed=seed) as sim:
+    with Simulator(net, minibatch_size=minibatch_size, seed=seed) as sim:
         x = np.outer(np.linspace(0, 1, batch_size),
-                     np.ones(step_blocks))[:, :, None]
+                     np.ones(n_steps))[:, :, None]
         y = np.outer(np.linspace(0, 1, batch_size),
-                     np.linspace(0, 1, step_blocks))[:, :, None]
+                     np.linspace(0, 1, n_steps))[:, :, None]
 
         sim.train({inp: x}, {p: y}, tf.train.GradientDescentOptimizer(2e-3),
                   n_epochs=2000)
 
         sim.check_gradients(sim.tensor_graph.losses[("mse", (p,))])
 
-        sim.run_steps(step_blocks, input_feeds={inp: x[:minibatch_size]})
+        sim.run_steps(n_steps, input_feeds={inp: x[:minibatch_size]})
 
     assert np.sqrt(np.mean((sim.data[p] - y[:minibatch_size]) ** 2)) < 0.05
 
 
-@pytest.mark.parametrize("unroll", (True, False))
+@pytest.mark.parametrize("unroll", (1, 2))
 def test_train_objective(Simulator, unroll, seed):
     minibatch_size = 1
-    step_blocks = 10
     n_hidden = 20
+    n_steps = 10
 
     with nengo.Network(seed=seed) as net:
         inp = nengo.Node([1])
@@ -243,10 +213,10 @@ def test_train_objective(Simulator, unroll, seed):
         nengo.Connection(inp, ens, synapse=0.01)
         p = nengo.Probe(ens)
 
-    with Simulator(net, minibatch_size=minibatch_size, step_blocks=step_blocks,
+    with Simulator(net, minibatch_size=minibatch_size,
                    unroll_simulation=unroll, seed=seed) as sim:
-        x = np.ones((minibatch_size, step_blocks, 1))
-        y = np.zeros((minibatch_size, step_blocks, 1))
+        x = np.ones((minibatch_size, n_steps, 1))
+        y = np.zeros((minibatch_size, n_steps, 1))
 
         def obj(output, target):
             return tf.reduce_mean((output[:, -1] - 0.5 - target[:, -1]) ** 2)
@@ -256,7 +226,7 @@ def test_train_objective(Simulator, unroll, seed):
 
         sim.check_gradients(sim.tensor_graph.losses[(obj, (p,))])
 
-        sim.run_steps(step_blocks, input_feeds={inp: x})
+        sim.run_steps(n_steps, input_feeds={inp: x})
 
         assert np.allclose(sim.data[p][:, -1], y[:, -1] + 0.5,
                            atol=1e-3)
@@ -264,7 +234,6 @@ def test_train_objective(Simulator, unroll, seed):
 
 def test_train_rmsprop(Simulator, seed):
     minibatch_size = 4
-    step_blocks = 1
     n_hidden = 5
 
     np.random.seed(seed)
@@ -285,8 +254,8 @@ def test_train_rmsprop(Simulator, seed):
 
         p = nengo.Probe(out.neurons)
 
-    with Simulator(net, minibatch_size=minibatch_size, step_blocks=step_blocks,
-                   unroll_simulation=True, seed=seed, device="/cpu:0") as sim:
+    with Simulator(net, minibatch_size=minibatch_size, unroll_simulation=1,
+                   seed=seed, device="/cpu:0") as sim:
         x = np.asarray([[[0, 0]], [[0, 1]], [[1, 0]], [[1, 1]]])
         y = np.asarray([[[0.1]], [[0.9]], [[0.9]], [[0.1]]])
 
@@ -302,7 +271,6 @@ def test_train_rmsprop(Simulator, seed):
 @pytest.mark.gpu
 def test_train_rmsprop_gpu(Simulator, seed):
     minibatch_size = 4
-    step_blocks = 1
     n_hidden = 5
 
     np.random.seed(seed)
@@ -323,8 +291,8 @@ def test_train_rmsprop_gpu(Simulator, seed):
 
         p = nengo.Probe(out.neurons)
 
-    with Simulator(net, minibatch_size=minibatch_size, step_blocks=step_blocks,
-                   unroll_simulation=True, seed=seed, device="/gpu:0") as sim:
+    with Simulator(net, minibatch_size=minibatch_size, unroll_simulation=1,
+                   seed=seed, device="/gpu:0") as sim:
         x = np.asarray([[[0, 0]], [[0, 1]], [[1, 0]], [[1, 1]]])
         y = np.asarray([[[0.1]], [[0.9]], [[0.9]], [[0.1]]])
 
@@ -341,8 +309,8 @@ def test_train_errors(Simulator):
         a = nengo.Node([0])
         p = nengo.Probe(a)
 
-    n_steps = 5
-    with Simulator(net, step_blocks=n_steps) as sim:
+    n_steps = 20
+    with Simulator(net) as sim:
         with pytest.raises(SimulationError):
             sim.train({a: np.ones((1, n_steps + 1, 1))},
                       {p: np.ones((1, n_steps, 1))}, None)
@@ -359,11 +327,8 @@ def test_train_errors(Simulator):
             sim.train({a: np.ones((1, n_steps, 1))},
                       {p: np.ones((1, n_steps, 2))}, None)
 
-    with Simulator(net, unroll_simulation=False, step_blocks=None) as sim:
-        with pytest.raises(SimulationError):
-            sim.train(None, None, None)
     with pytest.raises(SimulatorClosed):
-        sim.train(None, None, None)
+        sim.train({None: np.zeros((1, 1))}, None, None)
 
 
 def test_loss(Simulator):
@@ -374,16 +339,16 @@ def test_loss(Simulator):
         p = nengo.Probe(ens)
 
     n_steps = 20
-    with Simulator(net, step_blocks=n_steps) as sim:
-        sim.run_steps(20)
+    with Simulator(net) as sim:
+        sim.run_steps(n_steps)
         data = sim.data[p]
 
-        assert np.allclose(sim.loss({inp: np.ones((4, 20, 1))},
-                                    {p: np.zeros((4, 20, 1))}, "mse"),
+        assert np.allclose(sim.loss({inp: np.ones((4, n_steps, 1))},
+                                    {p: np.zeros((4, n_steps, 1))}, "mse"),
                            np.mean(data ** 2))
 
-        assert np.allclose(sim.loss({inp: np.ones((4, 20, 1))},
-                                    {p: np.zeros((4, 20, 1))},
+        assert np.allclose(sim.loss({inp: np.ones((4, n_steps, 1))},
+                                    {p: np.zeros((4, n_steps, 1))},
                                     objective=lambda x, y: tf.constant(2)),
                            2)
 
@@ -403,11 +368,8 @@ def test_loss(Simulator):
             sim.loss({inp: np.ones((1, n_steps, 1))},
                      {p: np.ones((1, n_steps, 2))}, None)
 
-    with Simulator(net, unroll_simulation=False, step_blocks=None) as sim:
-        with pytest.raises(SimulationError):
-            sim.loss(None, None, None)
     with pytest.raises(SimulatorClosed):
-        sim.loss(None, None, None)
+        sim.loss({None: np.zeros((1, 1))}, None, None)
 
 
 def test_generate_inputs(Simulator, seed):
@@ -418,7 +380,7 @@ def test_generate_inputs(Simulator, seed):
 
         p = [nengo.Probe(x) for x in inp]
 
-    with Simulator(net, minibatch_size=2, step_blocks=3) as sim:
+    with Simulator(net, minibatch_size=2, unroll_simulation=3) as sim:
         feed = sim._generate_inputs({inp[0]: np.zeros((2, 3, 1))}, 3)
 
         ph = [sim.tensor_graph.invariant_ph[x] for x in inp]
@@ -518,10 +480,8 @@ def test_model_passing(Simulator):
 
 
 @pytest.mark.gpu
-@pytest.mark.parametrize(
-    "device, unroll", itertools.product(["/cpu:0", "/gpu:0", None],
-                                        [True, False]))
-def test_devices(Simulator, device, unroll, seed):
+@pytest.mark.parametrize("device", ["/cpu:0", "/gpu:0", None])
+def test_devices(Simulator, device, seed):
     with nengo.Network(seed=seed) as net:
         inp = nengo.Node([0])
         ens = nengo.Ensemble(10, 1)
@@ -532,14 +492,12 @@ def test_devices(Simulator, device, unroll, seed):
         sim.run_steps(50)
         canonical = sim.data[p]
 
-    with Simulator(net, device=device, unroll_simulation=unroll,
-                   step_blocks=10) as sim:
+    with Simulator(net, device=device) as sim:
         sim.run_steps(50)
         assert np.allclose(sim.data[p], canonical)
 
 
-@pytest.mark.parametrize("unroll", (True, False))
-def test_side_effects(Simulator, unroll):
+def test_side_effects(Simulator):
     class MyFunc:
         x = 0
 
@@ -551,7 +509,7 @@ def test_side_effects(Simulator, unroll):
     with nengo.Network() as net:
         nengo.Node(output=func)
 
-    with Simulator(net, unroll_simulation=unroll, step_blocks=1) as sim:
+    with Simulator(net, unroll_simulation=1) as sim:
         sim.run_steps(10)
 
     assert func.x == 11
@@ -610,3 +568,11 @@ def test_probe_dict():
     assert len(a) == 2
     for x, y in zip(a, (0, 1)):
         assert x == y
+
+
+def test_deprecation(Simulator):
+    with pytest.warns(DeprecationWarning):
+        Simulator(None, step_blocks=1)
+
+    with pytest.warns(DeprecationWarning):
+        Simulator(None, unroll_simulation=True)